refactor(#3272): extract vec-access-exports from index.ts (refs #3182)
__vec_*/__dv_byte_*/__new_vec_f64 host-dispatch exports moved verbatim.
index.ts imports back the 4 finalize entry points + re-exports
reserveVecMethodHelper. Byte-identity IDENTICAL (39/39), tsc 0.

✓
